+++ include/bbbgpio/stream_direction.hpp
