+++ include/bbbgpio/stream_operation.hpp
