+++ src/CMakeLists.txt
@@ -16,4 +16,3 @@ install(TARGETS ${CMAKE_PROJECT_NAME}
 
 install(FILES ${HEADER_FILES}
         DESTINATION ${CMAKE_INSTALL_PREFIX}/include/${CMAKE_PROJECT_NAME})
-
